+++ dags/sentinel_2_nrt/env_cfg.py
@@ -2,12 +2,6 @@
 # Sentinel-2_nrt process env configs read from variables if environment specific
 """
 
-# OWS pod specific configuration
-OWS_CFG_PATH = "/env/config/ows_cfg.py"
-OWS_CFG_MOUNT_PATH = "/env/config"
-OWS_CFG_PATH = OWS_CFG_MOUNT_PATH + "/ows_cfg.py"
-OWS_CFG_IMAGEPATH = "/opt/dea-config/dev/services/wms/ows/ows_cfg.py"
-
 # TODO: The archive condition is for sentinel_2_nrt, set different condition for different products
 ARCHIVE_PRODUCTS = "s2a_nrt_granule s2b_nrt_granule"
 ARCHIVE_CONDITION = "[$(date -d '-365 day' +%F), $(date -d '-93 day' +%F)]"
@@ -46,23 +40,3 @@ UPDATE_EXTENT_PRODUCTS = (
 
 # batch indexing s3 paths
 S2_NRT_S3_PATHS = "s3://dea-public-data/L2/sentinel-2-nrt/S2MSIARD/**/ARD-METADATA.yaml"
-
-NODE_AFFINITY = {
-    "nodeAffinity": {
-        "requiredDuringSchedulingIgnoredDuringExecution": {
-            "nodeSelectorTerms": [
-                {
-                    "matchExpressions": [
-                        {
-                            "key": "nodetype",
-                            "operator": "In",
-                            "values": [
-                                "ondemand",
-                            ],
-                        }
-                    ]
-                }
-            ]
-        }
-    }
-}
