+++ dags/sentinel_2_nrt/sentinel_2_nrt_streamline_indexing.py
@@ -12,8 +12,8 @@ from airflow import DAG
 from airflow.kubernetes.secret import Secret
 from airflow.contrib.operators.kubernetes_pod_operator import KubernetesPodOperator
 
-from sentinel_2_nrt.images import INDEXER_IMAGE
-from env_var.infra import (
+from infra.images import INDEXER_IMAGE
+from infra.variables import (
     DB_DATABASE,
     DB_HOSTNAME,
     SECRET_ODC_WRITER_NAME,
@@ -24,8 +24,8 @@ from env_var.infra import (
 from sentinel_2_nrt.env_cfg import (
     INDEXING_PRODUCTS,
     PRODUCT_RECORD_PATHS,
-    NODE_AFFINITY,
 )
+from infra.podconfig import NODE_AFFINITY
 
 # DAG CONFIGURATION
 DEFAULT_ARGS = {
