+++ dags/sentinel_2_nrt/sentinel_2_nrt_update.py
@@ -10,9 +10,9 @@ from datetime import datetime, timedelta
 
 from airflow.kubernetes.secret import Secret
 from airflow.operators.subdag_operator import SubDagOperator
-from sentinel_2_nrt.subdag_ows_views import ows_update_extent_subdag
-from sentinel_2_nrt.subdag_explorer_summary import explorer_refresh_stats_subdag
-from env_var.infra import (
+from subdags.subdag_ows_views import ows_update_extent_subdag
+from subdags.subdag_explorer_summary import explorer_refresh_stats_subdag
+from infra.variables import (
     DB_DATABASE,
     DB_HOSTNAME,
     SECRET_AWS_NAME,
